Gets rid of require cycles
