+++ lib/ads/RewardedAd.js
@@ -16,7 +16,7 @@
  */
 
 import { isFunction, isString } from '../common';
-import googleAds from '../index';
+import googleAds from '../googleMobileAds';
 import validateAdRequestOptions from '../validateAdRequestOptions';
 import validateAdShowOptions from '../validateAdShowOptions';
 import MobileAd from './MobileAd';
